+++ spec/functional_spec_helper.rb
@@ -4,7 +4,17 @@ require 'database_cleaner'
 DatabaseCleaner.strategy = :truncation
 
 def functional_example?(metadata)
-  metadata[:file_path] =~ /spec\/functional/
+  metadata[:file_path] =~ %r{/spec/functional/}
+end
+
+def timeout(&blk)
+  Timeout.timeout(10, &blk)
+end
+
+def stub_tcp_connection(tcp_socket, ssl_socket, io_double)
+  allow_any_instance_of(Rpush::Daemon::TcpConnection).to receive_messages(connect_socket: [tcp_socket, ssl_socket])
+  allow_any_instance_of(Rpush::Daemon::TcpConnection).to receive_messages(setup_ssl_context: double.as_null_object)
+  stub_const('Rpush::Daemon::TcpConnection::IO', io_double)
 end
 
 RSpec.configure do |config|