+++ spec/spec_helper.rb
@@ -14,10 +14,7 @@ if !ENV['TRAVIS'] || (ENV['TRAVIS'] && ENV['QUALITY'] == 'true')
 end
 
 require 'timecop'
-
-if defined? JRUBY_VERSION
-  require 'activerecord-jdbc-adapter'
-end
+require 'activerecord-jdbc-adapter' if defined? JRUBY_VERSION
 
 require 'rpush'
 require 'rpush/daemon'
@@ -45,6 +42,7 @@ RPUSH_ROOT = '/tmp/rails_root'
 
 Rpush.configure do |config|
   config.client = client
+  config.log_level = ::Logger::Severity::DEBUG
 end
 
 RPUSH_CLIENT = Rpush.config.client